refactor(core/component): [Header] extract AccountSelector component
