+++ packages/next-starter/src/modules/account/components/index.ts
@@ -1,5 +1,6 @@
 export * from './Accountable';
 export * from './AccountForm';
 export * from './AccountList';
+export * from './AccountSelector';
 export * from './CreateAccountForm';
 export * from './UpdateAccountForm';
