+++ packages/next-starter/src/modules/core/components/Header.tsx
@@ -1,34 +1,20 @@
-import React, { useEffect } from 'react';
-import { useSelector } from 'react-redux';
-import { Avatar, Layout, Menu, Select, message } from 'antd';
+import React from 'react';
+import { Avatar, Layout, Menu } from 'antd';
 import { BarChart, Settings, Triangle } from 'react-feather';
 import Link from 'next/link';
 import { connect } from 'react-redux';
 import { bindActionCreators, Dispatch } from 'redux';
-import Router from 'next/router';
 
 import { wrapperActions, IWrapperPage, IStore } from '@onr/core';
-import { IAccount } from '@onr/account';
 
 import DashHeader from './styles/Header';
+import { AccountSelector } from '@onr/account';
 
 const { SubMenu } = Menu;
 
 const MainHeader = (props: IWrapperPage.IProps) => {
-  const { setOptionDrawer, setMobileDrawer, setAccountId } = props;
+  const { setOptionDrawer, setMobileDrawer } = props;
   const state = props;
-  const currentUser = useSelector((store: IStore) => store.authStore.currentUser);
-  const [accounts, setAccounts] = React.useState<IAccount[]>([]);
-
-  useEffect(() => {
-    if (currentUser.accounts) {
-      setAccounts(currentUser.accounts);
-
-      if (accounts[0] && !state.accountId) {
-        changeAccount(accounts[0].id);
-      }
-    }
-  }, [currentUser.accounts]);
 
   return (
     <DashHeader>
@@ -45,22 +31,7 @@ const MainHeader = (props: IWrapperPage.IProps) => {
           </a>
         </Link>
 
-        <Select
-          showSearch
-          style={{ width: 200 }}
-          optionFilterProp="children"
-          onChange={changeAccount}
-          value={state.accountId}
-          filterOption={(input, option) =>
-            option.props.children.toLowerCase().indexOf(input.toLowerCase()) >= 0
-          }
-        >
-          {accounts.map(account => (
-            <Select.Option key={account.id} value={account.id}>
-              {account.id} | {account.name}
-            </Select.Option>
-          ))}
-        </Select>
+        <AccountSelector />
 
         <span className="mr-auto" />
 
@@ -81,12 +52,6 @@ const MainHeader = (props: IWrapperPage.IProps) => {
       </Layout.Header>
     </DashHeader>
   );
-
-  function changeAccount(accountId: number) {
-    setAccountId(accountId);
-    Router.push('/');
-    message.info('Account has been changed');
-  }
 };
 
 const mapStateToProps = (state: IStore) => state.wrapper;
@@ -94,7 +59,6 @@ const mapStateToProps = (state: IStore) => state.wrapper;
 const mapDispatchToProps = (dispatch: Dispatch) => ({
   setOptionDrawer: bindActionCreators(wrapperActions.setOptionDrawer, dispatch),
   setMobileDrawer: bindActionCreators(wrapperActions.setMobileDrawer, dispatch),
-  setAccountId: bindActionCreators(wrapperActions.setAccountId, dispatch),
 });
 
 export const Header = connect(mapStateToProps, mapDispatchToProps)(MainHeader);
